Football Draft data import
This a simple example of a data import and conversion using T-SQL with some analysis mixed in. The data comes from a football mock draft made up of 96 teams and 1000's of players. Each player has 8 duplicates to be drafted. Data was captured using a web scraping script and imported via SQL server integration services